api_fastapi/requirements.txt
fastapi
uvicorn[standart]
python-dotenv
mysql-connector-python
python-dateutil